requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
qrcode
Pillow
WeasyPrint
gunicorn
psycopg2-binary
python-dotenv